+++ CHANGELOG.md
@@ -20,6 +20,7 @@ v0.15.0 (?? ??? 2018)
 - Save PNG alpha information when using remote images. @samsullivan #779
 - Fix parsing of `<w:br/>` tag. @troosan #1274
 - Bookmark are not writton as internal link in html writer @troosan #1263
+- It should be possible to add a Footnote in a ListItemRun @troosan #1287 #1287